Widgetize KPI progress statistics bar
